feat: atoms 분리 및 Notification Icon unmount될 때 Popover close
